requirements.txt
Django>=4.1.0,<4.2.0
djangorestframework>=3.14.0,<3.15
django-phonenumber-field>=7.2.0,<7.3
django-phonenumbers>=1.0.1,<1.1
PyYAML>=5.1.0,<5.2.0
uritemplate>=3.0.0,<3.1.0
markdown>3.0.0,<3.1.0
Pygments>=2.4.0,<2.5.0
django-filter>=1.0.1,<1.1.0
psycopg2>=2.9.9,<3.0
drf-spectacular>=0.27,<0.28
Pillow>=10.2.0,<10.3
